requirements.txt
pillow 
requests
fastapi
uvicorn
python-multipart